requirements.txt
flask
flask-cors
faiss-cpu
openai
python-dotenv
numpy
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client
